requirements.txt
Django>=5
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2-binary
python-dotenv
drf-yasg

